LayoutTests/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-in-strict-mode-wtih-checkbox.html

<!DOCTYPE html>
<div id="container"><input type=checkbox></div>
<div id="console"></div>
<script src="../../../../resources/js-test.js"></script>
<script>
var result = document.caretRangeFromPoint(40, 20);
shouldBeEqualToString('result.startContainer.tagName', 'DIV');
shouldBe('result.startOffset', '1');
shouldBe('result.startContainer', 'result.endContainer');
shouldBe('result.startOffset', 'result.endOffset');
if (window.testRunner)
    document.getElementById('container').outerHTML = '';
</script>
