third_party/blink/web_tests/platform/generic/fast/dom/Document/CaretRangeFromPoint/caretRangeFromPoint-in-strict-mode-wtih-checkbox-expected.txt

PASS result.startContainer.tagName is "DIV"
PASS result.startOffset is 1
PASS result.startContainer is result.endContainer
PASS result.startOffset is result.endOffset
PASS successfullyParsed is true

TEST COMPLETE
